+++ apps/examples/src/examples/HelloWorld.tsx
@@ -1,59 +1,8 @@
-import { PatchedMaterialMaster } from "@material-composer/patch-material"
 import { useControls } from "leva"
-import { initialModuleState, pipeModules } from "material-composer"
 import { Modules } from "material-composer-r3f"
-import {
-  ModuleRegistrationContext,
-  provideModuleRegistration
-} from "material-composer-r3f/src/moduleRegistration"
 import { Description } from "r3f-stage"
-import { useMemo } from "react"
-import { compileShader } from "shader-composer"
 import { useUniformUnit } from "shader-composer-r3f"
-import { patched } from "./lib/patched"
-
-const hasKey = <T extends object>(obj: T, k: keyof any): k is keyof T =>
-  k in obj
-
-export const composed = new Proxy(patched, {
-  get: (target, key) => {
-    if (!hasKey(target, key)) return
-
-    const Component = target[key]
-
-    return ({ children, ...props }: any) => {
-      const modules = provideModuleRegistration()
-
-      const shader = useMemo(() => {
-        /* Transform state with given modules. */
-        const { color, ...state } = pipeModules(
-          initialModuleState(),
-          ...(modules.list || [])
-        )
-
-        /* Construct a shader master unit */
-        const root = PatchedMaterialMaster({
-          ...state,
-          diffuseColor: color
-        })
-
-        console.log(root)
-
-        const [shader, meta] = compileShader(root)
-
-        return shader
-      }, [modules.version])
-
-      return (
-        <Component {...props} {...shader}>
-          <ModuleRegistrationContext.Provider value={modules}>
-            {children}
-          </ModuleRegistrationContext.Provider>
-        </Component>
-      )
-    }
-  }
-})
+import { composed } from "./lib/composed"
 
 export default function HelloWorld() {
   const controls = useControls({ mix: { value: 0.5, min: 0, max: 1 } })
